requirements.txt
Django==4.2.7
djangorestframework==3.14.0
python-dotenv==1.0.0
Pillow==10.1.0
django-cors-headers==4.3.0
django-filter==23.3
django-crispy-forms==2.1
crispy-bootstrap5==0.7
psycopg2-binary==2.9.9
gunicorn==21.2.0
whitenoise==6.6.0
django-environ==0.11.2
coverage==7.3.2
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
faker==20.0.3
django-debug-toolbar==4.2.0
django-extensions==3.2.3
django-allauth==0.58.2
django-rest-auth==0.9.5
django-storages==1.14.2
redis==5.0.1
celery==5.3.4
flower==2.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1
pytz==2023.3.post1 